requirements.txt
faker
python-dotenv
pillow
requests
torch
diffusers
transformers
accelerate
replicate==0.23.1
typing-extensions>=4.8.0